where2b_backend/where2b_core/requirements.txt
Django>=3.0,<4.0
djangorestframework>=3.0
djangorestframework-simplejwt==5.0.0
django-cors-headers==3.10.0 
django-filter==21.1  
drf-yasg==1.20.0
psycopg2-binary>=2.8
pandas==1.3.4
Pillow==8.4.0 
scipy==1.7.2 
